Create Solve Sudoku.java
Sudoku is one of the most popular puzzle games of all time. The goal of Sudoku is to fill a 9×9 grid with numbers so that each row, column and 3×3 section contain all of the digits between 1 and 9. As a logic puzzle, Sudoku is also an excellent brain game. If you play Sudoku daily, you will soon start to see improvements in your concentration and overall brain power.

So, given a partially filled 9×9 matrix i.e. mat[9][9], Your goal is to assign digits (from 1 to 9) to the empty cells so that every row, column, and sub-grid of size 3×3 contains exactly one instance of the digits from 1 to 9.

If the given sudoku is not solvable, or having wrong prefilled entries, you should return -1, otherwise on successful solution return 1.