Create Convert min Heap to max Heap .java
Given array representation of a min-heap convert it in a max-heap representation. For example, if array a = [13 15 19 16 18 120 110 112 118] the array must be modified as a = [120 118 110 112 18 19 13 15 16].

The array of numbers is given to you. Do not write the whole program, just complete the functions void modifyMintoMax(int array[], int n) which takes the array and total number of elements n as parameters and modify the array elements with heapify() to convert it in max-heap.

Note: Do not read any input from stdin/console. Just complete the functions provided. You can write more functions if required, but just above the given function.