Create Shortest path in a binary maze.java
Given a m x n binary matrix, find the length of the shortest path in the matrix from a given source cell to a destination cell. The path can only be created out of a cell if its value is 1.

If there is no path from a given source cell to a destination cell, return -1.

Note: At any point of time, you can either move horizontally or vertically in the four directions.