Create Interval scheduling Problem.java
Interval is defined as a span of time with start and end time. You are given N intervals with their start and finish times.

Two intervals can overlap each other if they have any time in common, e.g. Interval(5, 15) and Interval(10,20) are overlapping as time 10 to 15 is shared between them,

whereas Interval(5, 15) and Interval(20, 30) are non-overlapping as they do not have any common time.

Your task is to find the maximum number of intervals that can be scheduled, obviously they need to be non-overlapping.