Create Fractional Knapsack problem.java
The knapsack problem or rucksack problem is a problem in combinatorial optimization: Given a set of items, each with a weight and a value, determine the number of each item to include in a collection so that the total weight is less than or equal to a given limit and the total value is as large as possible.

In this problem we will solve a variant of it, in which the items may have a fractional decision i.e. you can pick the full item or partial depending on the capacity you have. Partial items are allowed to fill the bag at fullest. This is also known as fractional knapsack problem.

Given two integer arrays values[0..n-1] and weight[0..n-1] which represent values and weights associated with n items respectively. Also given an integer W which represents knapsack capacity, find out the maximum value subset of values[] such that sum of the weights of this subset is equal to W.