Create Activity Selection Problem.java
Activity is defined as a interval of time for its execution. Every activity has a start time and a finish time. You are given N activities with their start and finish times.

Select the maximum number of activities that can be performed by a single person, assuming that a person can only work on a single activity at a time.